Move aggregation to separate module
