+++ qa_eval/__init__.py
@@ -1,3 +1,4 @@
+from .aggregation import *
 from .evaluation import *
 from .steps import *
 from .steps.sparql import *
