+++ qa_eval/evaluation.py
@@ -1,7 +1,4 @@
-import json
 from collections import defaultdict
-from statistics import mean, median
-from typing import Any
 
 from .steps import get_steps_matches
 
@@ -135,119 +132,3 @@ def run_evaluation(
             })
             evaluation_results.append(eval_result)
     return evaluation_results
-
-
-def stats_for_series(values: list) -> dict[str, float]:
-    return {
-        "sum": sum(values),
-        "mean": mean(values) if values else 0,
-        "median": median(values) if values else 0,
-        "min": min(values) if values else 0,
-        "max": max(values) if values else 0,
-    }
-
-
-def compute_aggregates(samples: list[dict]) -> dict:
-    metrics = [
-        "answer_recall",
-        "answer_precision",
-        "answer_f1",
-        "steps_score",
-        "input_tokens",
-        "output_tokens",
-        "total_tokens",
-        "elapsed_sec"
-    ]
-    protected_metrics = [
-        "input_tokens",
-        "output_tokens",
-        "total_tokens",
-        "elapsed_sec"
-    ]
-    results_per_template = defaultdict(lambda: defaultdict(list))
-    number_of_samples_per_template_by_status = defaultdict(lambda: defaultdict(int))
-    steps_summary_per_template = defaultdict(lambda: defaultdict(lambda: defaultdict(int)))
-
-    templates_ids = set()
-    for sample in samples:
-        template_id = sample["template_id"]
-        templates_ids.add(template_id)
-
-        if "error" in sample:
-            number_of_samples_per_template_by_status[template_id]["error"] += 1
-            continue
-
-        number_of_samples_per_template_by_status[template_id]["success"] += 1
-
-        for metric in metrics:
-            value = sample.get(metric)
-            if value is not None:
-                results_per_template[template_id][metric].append(value)
-
-        seen = set()
-        for step in sample["actual_steps"]:
-            name = step["name"]
-            template_steps_summary = steps_summary_per_template[template_id]
-            template_steps_summary["total"][name] += 1
-            if step["status"] == "error":
-                template_steps_summary["errors"][name] += 1
-            if name not in seen:
-                seen.add(name)
-                template_steps_summary["once_per_sample"][name] += 1
-
-            if step["status"] != "error":
-                try:
-                    res = json.loads(step["output"])
-                    if "results" in res and "bindings" in res["results"]:
-                        if not res["results"]["bindings"]:
-                            template_steps_summary["empty_results"][name] += 1
-                except json.decoder.JSONDecodeError:
-                    pass
-
-    summary = {"per_template": {}}
-
-    for template_id in templates_ids:
-
-        template_summary: dict[str, Any] = {
-            "number_of_error_samples": number_of_samples_per_template_by_status[template_id]["error"],
-            "number_of_success_samples": number_of_samples_per_template_by_status[template_id]["success"],
-            "steps": {
-                k1: {k2: v2 for k2, v2 in v1.items()}
-                for k1, v1 in steps_summary_per_template[template_id].items()
-            },
-        }
-        for metric in metrics:
-            results_for_template = results_per_template[template_id]
-            series = results_for_template.get(metric, [])
-            if series or metric in protected_metrics:
-                template_summary[metric] = stats_for_series(series)
-
-        summary["per_template"][template_id] = template_summary
-
-    summary["micro"] = {
-        "number_of_error_samples": sum(
-            [values["error"] for values in number_of_samples_per_template_by_status.values()]),
-        "number_of_success_samples": sum(
-            [values["success"] for values in number_of_samples_per_template_by_status.values()]),
-    }
-    for metric in metrics:
-        series = [
-            i
-            for values in results_per_template.values()
-            for i in values[metric]
-            if values.get(metric) is not None
-        ]
-        if series or metric in protected_metrics:
-            summary["micro"][metric] = stats_for_series(series)
-
-    summary["macro"] = {}
-    for metric in metrics:
-        means = [
-            values[metric]["mean"]
-            for template_id, values in summary["per_template"].items()
-            if values.get(metric) is not None
-        ]
-        if means or metric in protected_metrics:
-            summary["macro"][metric] = {"mean": mean(means) if means else 0}
-
-    return summary
