+++ tests/test_evaluation.py
@@ -4,7 +4,7 @@ import jsonlines
 import yaml
 
 from qa_eval import (
-    compute_aggregations,
+    compute_aggregates,
     evaluate_steps,
     run_evaluation,
     stats_for_series,
@@ -21,7 +21,7 @@ def test_stats_for_series():
     }
 
 
-def test_run_evaluation_and_compute_aggregations():
+def test_run_evaluation_and_compute_aggregates():
     def get_chat_responses(path: Path) -> dict:
         responses = dict()
         with jsonlines.open(path, "r") as reader:
@@ -35,7 +35,7 @@ def test_run_evaluation_and_compute_aggregations():
     sample_chat_responses_path = Path(__file__).parent / "test_data" / "sample_chat_responses_1.jsonl"
 
     evaluation_results = run_evaluation(sample_reference_standard, get_chat_responses(sample_chat_responses_path))
-    aggregates = compute_aggregations(evaluation_results)
+    aggregates = compute_aggregates(evaluation_results)
     expected_evaluation_results = yaml.safe_load(
         (Path(__file__).parent / "test_data" / "sample_evaluation_per_question_1.yaml").read_text(encoding="utf-8")
     )
@@ -46,7 +46,7 @@ def test_run_evaluation_and_compute_aggregations():
     assert expected_aggregates == aggregates
 
 
-def test_run_evaluation_and_compute_aggregations_all_errors():
+def test_run_evaluation_and_compute_aggregates_all_errors():
     def get_chat_responses(path: Path) -> dict:
         responses = dict()
         with jsonlines.open(path, "r") as reader:
@@ -60,7 +60,7 @@ def test_run_evaluation_and_compute_aggregations_all_errors():
     sample_chat_responses_path = Path(__file__).parent / "test_data" / "sample_chat_responses_2.jsonl"
 
     evaluation_results = run_evaluation(sample_reference_standard, get_chat_responses(sample_chat_responses_path))
-    aggregates = compute_aggregations(evaluation_results)
+    aggregates = compute_aggregates(evaluation_results)
     expected_evaluation_results = yaml.safe_load(
         (Path(__file__).parent / "test_data" / "sample_evaluation_per_question_2.yaml").read_text(encoding="utf-8")
     )
